+++ PSFramework/internal/scripts/environment.ps1
@@ -86,3 +86,5 @@ if (($PSVersionTable.PSVersion.Major -ge 6) -and ($PSVersionTable.OS -notlike "*
 }
 
 if (-not ([PSFramework.Message.LogHost]::LoggingPath)) { [PSFramework.Message.LogHost]::LoggingPath = $script:path_Logging }
+
+[PSFramework.PSFCore.PSFCoreHost]::ModuleRoot = $script:ModuleRoot