+++ library/PSFramework/Validation/PsfValidateScriptAttribute.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Management.Automation;
+using static PSFramework.Extension.ScriptBlockExt;
 
 namespace PSFramework.Validation
 {
@@ -10,20 +11,43 @@ namespace PSFramework.Validation
     {
         /// <summary>
         /// Gets or sets the custom error message that is displayed to the user.
-        ///
-        /// The item being validated and the validating scriptblock is passed as the first and second
-        /// formatting argument.
-        ///
-        /// <example>
-        /// [ValidateScript("$_ % 2", ErrorMessage = "The item '{0}' did not pass validation of script '{1}'")]
-        /// </example>
         /// </summary>
-        public string ErrorMessage { get; set; }
+        public string ErrorMessage
+        {
+            get
+            {
+                if (!String.IsNullOrEmpty(ErrorString))
+                    return Localization.LocalizationHost.Read(ErrorString);
+                return _ErrorMessage;
+            }
+            set { _ErrorMessage = value; }
+        }
+        private string _ErrorMessage = "Cannot accept {0}, specify any of the following values: '{1}'";
+
+        /// <summary>
+        /// The stored localized string to use for error messages
+        /// </summary>
+        public string ErrorString;
 
         /// <summary>
         /// Gets the scriptblock to be used in the validation
         /// </summary>
-        public ScriptBlock ScriptBlock { get; }
+        public ScriptBlock ScriptBlock
+        {
+            get
+            {
+                if (!String.IsNullOrEmpty(ScriptBlockName))
+                    return Utility.UtilityHost.ScriptBlocks[ScriptBlockName].ScriptBlock;
+                return _ScriptBlock;
+            }
+            private set { _ScriptBlock = value; }
+        }
+        private ScriptBlock _ScriptBlock;
+
+        /// <summary>
+        /// Name of a stored scriptblock to use
+        /// </summary>
+        public string ScriptBlockName { get; private set; }
 
         /// <summary>
         /// Validates that each parameter argument matches the scriptblock
@@ -37,8 +61,8 @@ namespace PSFramework.Validation
                 throw new ValidationMetadataException("ArgumentIsEmpty", null);
             }
 
-            object result = ScriptBlock.Invoke(element);
-
+            object result = ScriptBlock.DoInvokeReturnAsIs(true, 2, element, null, null, new object[] { element });
+            
             if (!LanguagePrimitives.IsTrue(result))
             {
                 var errorMessageFormat = String.IsNullOrEmpty(ErrorMessage) ? "Error executing validation script: {0} against {{ {1} }}" : ErrorMessage;
@@ -54,6 +78,14 @@ namespace PSFramework.Validation
         public PsfValidateScriptAttribute(ScriptBlock ScriptBlock)
         {
             this.ScriptBlock = ScriptBlock ?? throw new ArgumentNullException("Need to specify a scriptblock!");
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the ValidateScriptBlockAttribute class
+        /// </summary>
+        public PsfValidateScriptAttribute(string ScriptBlockName)
+        {
+            this.ScriptBlockName = ScriptBlockName;
         }
     }
 }