+++ PSFramework/changelog.md
@@ -1,4 +1,20 @@
 ﻿# CHANGELOG
+## 0.10.29.160 : 2018-11-04
+ - New: Command ConvertTo-PSFHashtable converts objects into hashtables
+ - New: Command Get-PSFPipeline grants access to the current pipeline and all its works.
+ - New: Command Get-PSFScriptblock retrieves scriptblocks from a static dictionary
+ - New: Command Set-PSFScriptblock stores scriptblocks in a static dictionary
+ - New: Command Get-PSFLocalizedString retrieves localied versions of strings
+ - New: Command Import-PSFLocalizedString imports localized strings into the strings store
+ - New: Logging Provider for gelf / graylog
+ - Upd: PsfValidateScript can now consume stored scriptblocks
+ - Upd: PsfValidateScript will now understand both $_ and $args[0]
+ - Upd: PsfValidateSet now supports localized strings using the `ErrorString` property.
+ - Upd: PsfValidateScript now supports localized strings using the `ErrorString` property.
+ - Upd: Logging runspace now loads the same copy of PSFramework that spawned it (#238)
+ - Fix: PsfValidateSet fails on completion scriptblock with whitespace value
+ - Fix: Get-PSFConfig will show bad value in default table. Correct data still stored (#243)
+
 ## 0.10.28.144 : 2018-10-28
  - Upd: Module Architecture update
  - Upd: Linked online help for commands (by Andrew Pla)